Refactor recipes resolution
Summary:
This diffs refactor the recipes resolution. Previously we defined an entity Recipe that required to know unrelated params in order to resolve its metadata and logic.
We splitted that functionality in two different classes RecipeMetadataProvider and RecipeLogicProvider so the ComponentProvider, in order to return a recipe uses this two Provider and gather the result.
We have refactored the tests as well and increased the coverage of the LogicProvider comparing with the previous state.
Apart from the above we implemented a fallback to be able to use Anvil and local files in the MetadataProvider. This feature is not covered by the tests since it is temporary behaviour until we finally merge the required changes in the Anvil project.

Reviewers: jfuentes, james

Differential Revision: http://phabricator.bitnami.com:8080/D12121