Recursive search for excluded files
Summary:
Previously the exlude paths were just applied to the root of the component prefix.
This differ from the `exclude` option of `tar` that excludes the file even if it is in a deeper directory.

Note that we need to exclude files per prefix in just one `tar` call so we still need to use absolute paths when excluding.

Reviewers: jfuentes

Differential Revision: http://phabricator.bitnami.com:8080/D14080