Report system packages in the build summary
Summary:
This diff includes the changes needed to report in the final summary the list of runtime packages required for the compiled binaries.

For doing so we use Distro classes that defines the particularities of each distro. The methods 'updateCommand' and 'getInstallCommands' are not used yet but they are a preparation for installing build dependencies.

Reviewers: jfuentes, juanjo

Reviewed By: jfuentes

Differential Revision: http://phabricator.bitnami.com:8080/D14899