Use tar library instead tarballUtils when compressing artifacts
Summary:
The problem with the tar method of tarball utils is that if the list of files to compress is too big it throws an error.
We reduced the amount of files passed to tarballUtils.tar but that slowed down the process a lot.

Instead of that we are now using the 'tar' library for node since it allowes any amount of files as arguments. To adapt
the code to 'tar' we needed:
 - Use relative paths when calling for compression
 - Exclude the files beforehand. `tar.c` doesn't support to exclude files.

Reviewers: jfuentes

Differential Revision: http://phabricator.bitnami.com:8080/D16359